requirements.txt
discord.py>=2.0
requests>=2.28
flask>=2.0
python-dotenv>=0.19
discord.py[voice]==2.3.2